+++ src/components/main_view.js
@@ -6,12 +6,21 @@ import Chore from "./chore";
 import NewTask from "./NewTask";
 import TextFieldForm from "./TextFieldForm";
 import ExtendedLogger from "./ExtendedLogger";
+import UserSearchBox from "./UserSearchBox";
+import Dropdown from "./Dropdown";
+import * as PropTypes from "prop-types";
 
 export default class MainView extends Component {
+  propTypes = {
+    API: PropTypes.object,
+    list: PropTypes.object,
+  }
+
   constructor(props) {
     super(props);
     
-    let listItems = props.list === null ? [] : props.list.items
+    let listItems = props.list === null ? [] : props.list.items;
+    let sharedWith = props.list === null ? [] : props.list.shared_with;
     
     this.state = {
       editName: false,
@@ -20,6 +29,12 @@ export default class MainView extends Component {
         lock: false,
         error: false
       },
+      sharing: {
+        display: false,
+        lock: false,
+        error: false,
+      },
+      showSharedWith: false,
       selectedChore: null,
       choreDetails: {logs: []},
       /**
@@ -30,6 +45,7 @@ export default class MainView extends Component {
        * @params {[object]}
        */
       listItems: listItems,
+      sharedWith: sharedWith,
       editingTask: null,
     };
 
@@ -41,9 +57,54 @@ export default class MainView extends Component {
     this.choreDetails = this.choreDetails.bind(this);
     this.toggleEditing = this.toggleEditing.bind(this);
     this.updateName = this.updateName.bind(this);
+    this.shareList = this.shareList.bind(this);
+    this.unshareList = this.unshareList.bind(this);
+    this.toggleSharing = this.toggleSharing.bind(this);
+    this.toggleSharedWith = this.toggleSharedWith.bind(this);
   }
 
   // List actions
+  toggleSharing() {
+    let newObject = {...this.state.sharing, display: !this.state.sharing.display};
+    console.log("Display sharing window:", newObject.display);
+    this.setState({sharing: newObject});
+  }
+
+  toggleSharedWith() {
+    this.setState({showSharedWith: !this.state.showSharedWith});
+  }
+
+  /**
+   *
+   * @param profile {object}
+   * @returns {Promise<void>}
+   */
+  async shareList(profile) {
+    console.log(this.state.sharedWith);
+    if (this.state.sharedWith.findIndex(x => x.id === profile.id) > -1) {
+      console.log(`Already shared with user ${profile.id}`);
+      return;
+    }
+    console.log(`Sharing with user ${profile.id}`);
+    try {
+      await this.props.API.shareChecklist(this.props.list.id, profile.id);
+    } catch (e) {
+      console.error(e);
+      return;
+    }
+    this.setState({sharedWith: [...this.state.sharedWith, profile]})
+  }
+
+  async unshareList(pk) {
+    try {
+      await this.props.API.unshareChecklist(this.props.list.id, pk);
+    } catch (e) {
+      console.error(e);
+      return;
+    }
+    this.setState({sharedWith: this.state.sharedWith.filter(profile => profile.id !== pk)})
+  }
+
   toggleEditing() {
     this.setState({editName: !this.state.editName});
   }
@@ -59,15 +120,15 @@ export default class MainView extends Component {
     let taskName = document.querySelector("#chore-input").value;
     let frequency = document.querySelector("#chore-frequency").value;
     try {
-      let response = await this.props.addTask(taskName, frequency);
-      if (response.status === 200) {
+      let success = await this.props.addTask(taskName, frequency);
+      if (success) {
         this.toggleNewTask();
         return;
       }
     } catch {
 
     }
-    this.setState({addTask: {display: true, lock: false, error: "Det oppstod ein feil. Prøv igjen."}})
+    this.setState({addTask: {display: true, lock: false, error: "Det oppstod ein feil. Prøv igjen."}});
   }
 
   toggleNewTask() {
@@ -88,7 +149,13 @@ export default class MainView extends Component {
   }
 
   async logChore(pk, note="", dtg=null) {
-    let json = await this.props.logChore(pk, note, dtg);
+    if (dtg === null) {
+      console.log(`Logging chore ${pk} with note ${note} now...`);
+      dtg = new Date();
+    } else {
+      console.log(`Logging chore ${pk} with note ${note} at ${dtg.toJSON()}...`);
+    }
+    let json = await this.props.API.logTask(pk, note, dtg);
     let choreDetails = this.state.selectedChore === pk ? json : this.state.choreDetails;
     let extendedLogger = this.state.extendedLogger === pk ? null : this.state.extendedLogger;
     let listItemsCopy = this.state.listItems.filter(x => {
@@ -131,18 +198,26 @@ export default class MainView extends Component {
       return;
     }
 
-    let chore = await this.props.getChore(pk);
-    console.log(chore);
+    let task;
+
+    try {
+      task = await this.props.API.task(pk);
+    } catch (e) {
+      console.error(e);
+      return;
+    }
+
+    console.log(task);
     this.setState(
       {
         selectedChore: pk,
-        choreDetails: chore,
+        choreDetails: task,
       }
     );
   }
   
   async deleteLog(pk) {
-    let json = await this.props.deleteLog(pk);
+    let json = await this.props.API.deleteLog(pk);
     let listItemsCopy = this.state.listItems.filter(x => {
       if (x.id === json.id) {
         x.last_logged = json.last_logged;
@@ -161,13 +236,24 @@ export default class MainView extends Component {
   componentDidUpdate(prevProps, prevState, snapshot) {
     console.log(this.props.list);
     let listItems = [];
+    let sharedWith = [];
     if (this.props.list !== null) {
       if (this.props.list.items !== undefined) {
         if (prevProps.list === null || prevProps.list.items !== this.props.list.items) {
           listItems = this.props.list.items.map(x => x);
-          this.setState({listItems: listItems});
         }
       }
+      if (this.props.list.shared_with !== undefined) {
+        if (prevProps.list === null || prevProps.list.shared_with !== this.props.list.shared_with) {
+          sharedWith = this.props.list.shared_with.map(x => x);
+        }
+      }
+    }
+    if (listItems.length > 0 || sharedWith.length > 0) {
+      this.setState({
+        listItems: listItems,
+        sharedWith: sharedWith,
+      })
     }
   }
 
@@ -175,6 +261,8 @@ export default class MainView extends Component {
     if (this.props.list === null) {
       return <div className="col-md-9 col-lg-10">Du har ikkje valt noka liste</div>;
     }
+
+    console.log(this.state.listItems);
     
     let table = this.state.listItems.length === 0
       ? <span>Her er det ingen oppgåver. Prøv å leggje til ei.</span>
@@ -206,34 +294,89 @@ export default class MainView extends Component {
           {extendedLogger}
           {details}
         </>
-        })
+        });
 
     return <div className="col-md-9 col-lg-10">
-      { !this.state.editName
-        ? <h3>{ this.props.list.name }</h3>
-        : <TextFieldForm value={this.props.list.name} completion={(name) => {this.updateName(name)}}
-                         id={"list-name"} caption={"Lagre"}
-          /> }
+      <div className={"row"}>
+        <div className={"col"}>
+        { !this.state.editName
+          ? <h3>{this.props.list.name}</h3>
+          : <TextFieldForm value={this.props.list.name}
+                           completion={(name) => {this.updateName(name)}}
+                           id={"list-name"} caption={"Lagre"}
+            />
+        }
+        </div>
+        <div className={"col text-end"}>
+          <div className="btn-group">
+            <div className={"btn text-muted" + (this.state.sharedWith.length > 0 ? " dropdown-toggle" : "")}
+                 onClick={this.toggleSharedWith}
+            >
+              {
+              this.state.sharedWith.length > 0
+                ? `Delt med ${this.state.sharedWith.length}`
+                : "Ikkje delt med nokon"
+            }
+            </div>
+            <Dropdown visible={this.state.showSharedWith} xOffset={-34}>
+              {
+                this.state.sharedWith.map(profile => <li className={"dropdown-item"}>
+                  <div className={"row"}>
+                    <div className={"col"}>{profile.name}</div>
+                    { ((this.props.profile.id === profile.id) ||
+                        (this.props.profile.id === this.props.list.owner.id)) &&
+                    <div className={"col text-end"}>
+                      <i className={"bi bi-trash3"}
+                         onClick={() => {this.unshareList(profile.id)}}
+                      />
+                    </div> }
+                  </div>
+                </li>)
+              }
+            </Dropdown>
+          </div>
+        </div>
+      </div>
+
       <Button key={"new-chore"}
               icon={"plus-circle"} caption={"Ny oppgåve"}
               completion={() => this.toggleNewTask()}
               classNames={"btn-primary"}
       />
-      <Button key={"edit-list"}
-              icon={"pencil"} caption={"Rediger"}
-              completion={() => {this.toggleEditing()}}
-              classNames={"btn-primary"}
-      />
-      <Button key={"delete-list"}
-              icon={"trash3"} caption={"Slett lista"}
-              completion={() => {this.props.deleteList()}}
-              classNames={"btn-danger"}
-      />
+      { this.props.list.owner.id === this.props.profile.id
+      && <>
+        <div className={"btn-group"}>
+          <Button key={"share-list"}
+                  icon={"box-arrow-up"} caption={"Del"}
+                  completion={() => this.toggleSharing()}
+                  classNames={"btn-primary rounded"}
+          />
+          <Dropdown visible={this.state.sharing.display}>
+            <UserSearchBox API={this.props.API} completion={this.shareList}
+                           sharedWith={this.state.sharedWith}/>
+          </Dropdown>
+        </div>
+        <Button key={"edit-list"}
+        icon={"pencil"} caption={"Rediger"}
+        completion={() => {
+        this.toggleEditing()
+      }}
+        classNames={"btn-primary"}
+        />
+        <Button key={"delete-list"}
+        icon={"trash3"} caption={"Slett lista"}
+        completion={() => {
+        this.props.deleteList()
+      }}
+        classNames={"btn-danger"}
+        />
+        </>
+      }
       <div className={"d-grid g-2"}>
-      { this.state.addTask.display
-        ? <NewTask state={this.state.addTask} completion={() => this.addTask()}/>
-        : null }
-      { table }
+        {this.state.addTask.display
+          ? <NewTask state={this.state.addTask} completion={() => this.addTask()}/>
+          : null}
+        {table}
       </div>
     </div>;
   }